HSM-14-19 decomposition cont.: carve CompanionMesh + QueuePresence (app 4,900 -> 3,107)
Two more verbatim carves out of MeetingCaptureApp.swift into App/MeetingCapture/:
- CompanionMesh.swift (1,294) — the Agent Desk (HSM-15-08), the dictate-into-your-Mac peer
  (Phase 13/15), and LAN mesh discovery/pairing (NWBrowser). #if/#endif balanced.
- QueuePresence.swift (498) — the app-wide run-queue transparency layer (RunQueueStore/QueueHUD)
  + the presence-nudge overlay (HSM-14-15).

Device-arch xcodebuild BUILD SUCCEEDED after each. MeetingCaptureApp.swift is now 3,107
(8,042 at the decomposition's start — eight modules carved, the core file down 61%).
